create main-rescan, organize methods for rescanning more coherently
